+++ lib/mix/tasks/phoenix/pow.phoenix.install.ex
@@ -23,7 +23,6 @@ defmodule Mix.Tasks.Pow.Phoenix.Install do
   """
   use Mix.Task
 
-  alias Pow.Config
   alias Mix.{Pow, Pow.Phoenix}
   alias Mix.Tasks.Pow.Extension.Phoenix.Gen.Templates, as: PhoenixExtensionTemplatesTask
   alias Mix.Tasks.Pow.Phoenix.Gen.Templates, as: PhoenixTemplatesTask
@@ -43,13 +42,131 @@ defmodule Mix.Tasks.Pow.Phoenix.Install do
     args
     |> Pow.parse_options(@switches, @default_opts)
     |> parse_structure()
+    |> print_shell_instructions(schema_opts)
     |> maybe_run_gen_templates(args)
     |> maybe_run_extensions_gen_templates(args)
-    |> print_shell_instructions(schema_opts)
   end
 
   defp parse_structure({config, _parsed, _invalid}) do
     Map.put(config, :structure, Phoenix.parse_structure(config))
+  end
+
+  defp print_shell_instructions(%{structure: structure} = config, schema_opts) do
+    [
+      config_file_injection(structure, schema_opts),
+      phoenix_endpoint_file_injection(structure),
+      phoenix_router_file_injection(structure)
+    ]
+    |> Pow.inject_files()
+    |> case do
+      :ok ->
+        Mix.shell().info("Pow has been installed in your Phoenix app!")
+        config
+
+      :error ->
+        Mix.raise "Couldn't install Pow! Did you run this inside your Phoenix app?"
+    end
+  end
+
+  defp config_file_injection(structure, schema_opts) do
+    file = Path.expand(Keyword.fetch!(Mix.Project.config(), :config_path))
+
+    content =
+      """
+      config #{inspect(structure.web_app)}, :pow,
+        user: #{inspect(structure.context_base)}.#{schema_opts.schema_name},
+        repo: #{inspect(structure.context_base)}.Repo
+      """
+
+    %{
+      file: file,
+      injections: [%{
+        content: content,
+        test: "config #{inspect(structure.web_app)}, :pow",
+        needle: "import_config",
+        prepend: true
+      }],
+      instructions:
+        """
+        Append this to #{Path.relative_to_cwd(file)}:
+
+        #{content}
+        """
+    }
+  end
+
+  defp phoenix_endpoint_file_injection(structure) do
+    file = Path.expand("#{structure.web_prefix}/endpoint.ex")
+    content = "  plug Pow.Plug.Session, otp_app: #{inspect(structure.web_app)}"
+
+    %{
+      file: file,
+      injections: [%{
+        content: content,
+        test: "plug Pow.Plug.Session",
+        needle: "plug Plug.Session"
+      }],
+      instructions:
+        """
+        Add the `Pow.Plug.Session` plug to #{Path.relative_to_cwd(file)} after the `Plug.Session` plug:
+
+        defmodule #{inspect(structure.web_module)}.Endpoint do
+          use Phoenix.Endpoint, otp_app: #{inspect(structure.web_app)}
+
+          # ...
+
+          plug Plug.Session, @session_options
+        #{content}
+          plug #{inspect(structure.web_module)}.Router
+        end
+        """
+    }
+  end
+
+  defp phoenix_router_file_injection(structure) do
+    file = Path.expand("#{structure.web_prefix}/router.ex")
+
+    router_use_content = "  use Pow.Phoenix.Router"
+    router_scope_content =
+      """
+        scope "/" do
+          pipe_through :browser
+
+          pow_routes()
+        end
+      """
+
+    %{
+      file: file,
+      injections: [
+        %{
+          content: router_use_content,
+          test: "use Pow.Phoenix.Router",
+          needle: "use #{inspect(structure.web_module)}, :router"
+        },
+        %{
+          content: router_scope_content,
+          test: "pow_routes()",
+          needle: "scope ",
+          prepend: true,
+        }
+      ],
+      instructions:
+        """
+        Update `#{Path.relative_to_cwd(file)}` with the Pow routes:
+
+        defmodule #{inspect(structure.web_module)}.Router do
+          use #{inspect(structure.web_module)}, :router
+        #{router_use_content}
+
+          # ... pipelines
+
+        #{router_scope_content}
+
+          # ... routes
+        end
+        """
+    }
   end
 
   defp maybe_run_gen_templates(%{templates: true} = config, args) do
@@ -65,61 +182,4 @@ defmodule Mix.Tasks.Pow.Phoenix.Install do
     config
   end
   defp maybe_run_extensions_gen_templates(config, _args), do: config
-
-  defp print_shell_instructions(%{structure: %{context_base: context_base, web_module: web_base, web_prefix: web_prefix, web_app: web_app}}, %{schema_name: schema_name}) do
-    case config_set?(web_app) do
-      true ->
-        :ok
-
-      false ->
-        Mix.shell().info(
-          """
-          Pow has been installed in your phoenix app!
-
-          There are three files you'll need to configure first before you can use Pow.
-
-          First, append this to `config/config.exs`:
-
-          config #{inspect(web_app)}, :pow,
-            user: #{inspect(context_base)}.#{schema_name},
-            repo: #{inspect(context_base)}.Repo
-
-          Next, add `Pow.Plug.Session` plug to `#{web_prefix}/endpoint.ex` after `plug Plug.Session`:
-
-          defmodule #{inspect(web_base)}.Endpoint do
-            use Phoenix.Endpoint, otp_app: #{inspect(web_app)}
-
-            # ...
-
-            plug Plug.Session, @session_options
-            plug Pow.Plug.Session, otp_app: #{inspect(web_app)}
-            plug #{inspect(web_base)}.Router
-          end
-
-          Last, update `#{web_prefix}/router.ex` with the Pow routes:
-
-          defmodule #{inspect(web_base)}.Router do
-            use #{inspect(web_base)}, :router
-            use Pow.Phoenix.Router
-
-            # ... pipelines
-
-            scope "/" do
-              pipe_through :browser
-
-              pow_routes()
-            end
-
-            # ... routes
-          end
-          """)
-    end
-  end
-
-  defp config_set?(web_app) do
-    user = Config.get([otp_app: web_app], :user)
-    repo = Config.get([otp_app: web_app], :repo)
-
-    not is_nil(user) && not is_nil(repo)
-  end
 end