+++ test/mix/tasks/extension/ecto/pow.extension.ecto.gen.migrations_test.exs
@@ -30,26 +30,18 @@ defmodule Mix.Tasks.Pow.Extension.Ecto.Gen.MigrationsTest do
 
   defmodule Repo do
     def __adapter__, do: true
-    def config, do: [priv: "tmp/#{inspect(Migrations)}", otp_app: :pow]
+    def config, do: [priv: "./", otp_app: :pow]
   end
 
   use Pow.Test.Mix.TestCase
 
   @extension       ExtensionMock
   @extension_name  "MixTasksPowExtensionEctoGenMigrationsTestExtensionMock"
-  @tmp_path        Path.join(["tmp", inspect(Migrations)])
-  @migrations_path Path.join([@tmp_path, "migrations"])
+  @migrations_path "migrations"
   @options         ["-r", inspect(Repo), "--extension", @extension]
 
-  setup do
-    File.rm_rf!(@tmp_path)
-    File.mkdir_p!(@tmp_path)
-
-    :ok
-  end
-
-  test "generates migrations" do
-    File.cd!(@tmp_path, fn ->
+  test "generates migrations", context do
+    File.cd!(context.tmp_path, fn ->
       Migrations.run(@options)
 
       assert [migration_file] = File.ls!(@migrations_path)
@@ -65,26 +57,26 @@ defmodule Mix.Tasks.Pow.Extension.Ecto.Gen.MigrationsTest do
     end)
   end
 
-  test "warns if no extensions" do
-    File.cd!(@tmp_path, fn ->
+  test "warns if no extensions", context do
+    File.cd!(context.tmp_path, fn ->
       Migrations.run(["-r", inspect(Repo)])
 
       assert_received {:mix_shell, :error, ["No extensions was provided as arguments, or found in `config :pow, :pow` configuration."]}
     end)
   end
 
-  test "warns no migration file" do
-    File.cd!(@tmp_path, fn ->
+  test "warns no migration file", context do
+    File.cd!(context.tmp_path, fn ->
       Migrations.run(["-r", inspect(Repo), "--extension", "PowResetPassword"])
 
       assert_received {:mix_shell, :info, ["Notice: No migration file will be generated for PowResetPassword as this extension doesn't require any migrations."]}
     end)
   end
 
-  test "generates with `:binary_id`" do
+  test "generates with `:binary_id`", context do
     options = @options ++ ~w(--binary-id)
 
-    File.cd!(@tmp_path, fn ->
+    File.cd!(context.tmp_path, fn ->
       Migrations.run(options)
 
       assert [migration_file] = File.ls!(@migrations_path)
@@ -104,8 +96,8 @@ defmodule Mix.Tasks.Pow.Extension.Ecto.Gen.MigrationsTest do
       end)
     end
 
-    test "generates migrations" do
-      File.cd!(@tmp_path, fn ->
+    test "generates migrations", context do
+      File.cd!(context.tmp_path, fn ->
         Application.put_env(:pow, :pow, extensions: [@extension])
         Migrations.run(["-r", inspect(Repo)])
 
@@ -114,23 +106,23 @@ defmodule Mix.Tasks.Pow.Extension.Ecto.Gen.MigrationsTest do
     end
   end
 
-  test "doesn't make duplicate migrations" do
+  test "doesn't make duplicate migrations", context do
     options = @options ++ ["--extension", @extension]
 
-    File.cd!(@tmp_path, fn ->
+    File.cd!(context.tmp_path, fn ->
       assert_raise Mix.Error, "migration can't be created, there is already a migration file with name Add#{@extension_name}ToUsers.", fn ->
         Migrations.run(options)
       end
     end)
   end
 
-  test "generates with `:generators` config" do
+  test "generates with `:generators` config", context do
     Application.put_env(:pow, :generators, binary_id: true)
     on_exit(fn ->
       Application.delete_env(:pow, :generators)
     end)
 
-    File.cd!(@tmp_path, fn ->
+    File.cd!(context.tmp_path, fn ->
       Migrations.run(@options)
 
       assert [migration_file] = File.ls!(@migrations_path)