+++ test/mix/tasks/pow.install_test.exs
@@ -3,36 +3,27 @@ defmodule Mix.Tasks.Pow.InstallTest do
 
   alias Mix.Tasks.Pow.Install
 
-  @tmp_path Path.join(["tmp", inspect(Install)])
-
-  setup do
-    File.rm_rf!(@tmp_path)
-    File.mkdir_p!(@tmp_path)
-
-    :ok
-  end
-
-  test "generates files" do
-    File.cd!(@tmp_path, fn ->
+  test "generates files", context do
+    File.cd!(context.tmp_path, fn ->
       Install.run([])
 
       assert File.ls!("lib/pow/users") == ["user.ex"]
     end)
   end
 
-  test "with schema name and table" do
+  test "with schema name and table", context do
     options = ~w(Accounts.Organization users)
 
-    File.cd!(@tmp_path, fn ->
+    File.cd!(context.tmp_path, fn ->
       Install.run(options)
 
-      assert_received {:mix_shell, :info, ["Pow has been installed in your phoenix app!" <> msg]}
-      assert msg =~ "user: Pow.Accounts.Organization"
+      assert_received {:mix_shell, :info, ["Pow has been installed in your Phoenix app!"]}
+      assert File.read!(context.paths.config_path) =~ "user: Pow.Accounts.Organization"
     end)
   end
 
-  test "raises error in umbrella app" do
-    File.cd!(@tmp_path, fn ->
+  test "raises error in umbrella app", context do
+    File.cd!(context.tmp_path, fn ->
       File.write!("mix.exs", """
       defmodule Umbrella.MixProject do
         use Mix.Project
@@ -51,8 +42,8 @@ defmodule Mix.Tasks.Pow.InstallTest do
     end)
   end
 
-  test "raises error on invalid schema name or table" do
-    File.cd!(@tmp_path, fn ->
+  test "raises error on invalid schema name or table", context do
+    File.cd!(context.tmp_path, fn ->
       assert_raise Mix.Error, ~r/Invalid arguments/, fn ->
         Install.run(~w(Users.User))
       end